// backend/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const TaxonomyRank = {
  kingdom: 'kingdom',
  phylum: 'phylum',
  class_: 'class_',
  order: 'order',
  family: 'family',
  genus: 'genus',
  species: 'species'
} as const

export type TaxonomyRank = (typeof TaxonomyRank)[keyof typeof TaxonomyRank]


export const NomenclatureStatus = {
  accepted: 'accepted',
  synonym: 'synonym'
} as const

export type NomenclatureStatus = (typeof NomenclatureStatus)[keyof typeof NomenclatureStatus]


export const RegionType = {
  continent: 'continent',
  country: 'country',
  region: 'region',
  province: 'province',
  locality: 'locality'
} as const

export type RegionType = (typeof RegionType)[keyof typeof RegionType]


export const UserRole = {
  admin: 'admin',
  editor: 'editor',
  user: 'user'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const ContributionType = {
  suggest_edit: 'suggest_edit',
  add_image: 'add_image',
  add_plant: 'add_plant',
  report_error: 'report_error'
} as const

export type ContributionType = (typeof ContributionType)[keyof typeof ContributionType]


export const ContributionStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected'
} as const

export type ContributionStatus = (typeof ContributionStatus)[keyof typeof ContributionStatus]


export const Observability = {
  easy: 'easy',
  moderate: 'moderate',
  hard: 'hard'
} as const

export type Observability = (typeof Observability)[keyof typeof Observability]


export const TraitDataType = {
  option: 'option',
  number: 'number',
  boolean: 'boolean',
  text: 'text'
} as const

export type TraitDataType = (typeof TraitDataType)[keyof typeof TraitDataType]


export const PlantGroup = {
  angiosperm: 'angiosperm',
  gymnosperm: 'gymnosperm',
  fern: 'fern'
} as const

export type PlantGroup = (typeof PlantGroup)[keyof typeof PlantGroup]